// src/shared/dataprovider/api/model/eODAgentComms.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface EODAgentComms { 
    Agent?: string;
    TotalGross?: number;
    DepositGross?: number;
    WithdrawalGross?: number;
    EODComms?: number;
    WaletBalance?: number;
}