// src/shared/dataprovider/api/model/eODAgentCommsSummary.ts
/**
 * DynastyGateway API
 * API Connections to DPay
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EODAgentComms } from './eODAgentComms';
import { EODAgentCommsGrandTotals } from './eODAgentCommsGrandTotals';

export interface EODAgentCommsSummary { 
    GrandTotals?: EODAgentCommsGrandTotals;
    Reports?: Array<EODAgentComms>;
    TotalRecordCount?: number;
    Pagesize?: number;
    PageNumber?: number;
}