// src/shared/dataprovider/api/model/eODTransactionSummaryReport.ts
/**
 * DynastyGateway API
 * API Connections to DPay
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TransactionEODGrandTotals } from './transactionEODGrandTotals';
import { TransactionReportEOD } from './transactionReportEOD';

export interface EODTransactionSummaryReport { 
    GrandTotals?: TransactionEODGrandTotals;
    Reports?: Array<TransactionReportEOD>;
    TotalRecordCount?: number;
    Pagesize?: number;
    PageNumber?: number;
}