// src/shared/dataprovider/api/model/transactionReportEOD.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TransactionTypeEnum } from './transactionTypeEnum';

export interface TransactionReportEOD { 
    Client?: string;
    Merchant?: string;
    Type?: TransactionTypeEnum;
    Channel?: string;
    Count?: number;
    Gross?: number;
    Fees?: number;
    Net?: number;
    TotalProfit?: number;
}