// src/shared/dataprovider/api/model/vendorGrandTotals.ts
/**
 * DynastyGateway API
 * API Connections to DPay
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface VendorGrandTotals { 
    GrandTotalFees?: number;
    GrandTotalGross?: number;
    GrandTotalACNet?: number;
    GrandTotalOTNet?: number;
    GrandTotalNet?: number;
    GrandTotalProfit?: number;
}