// src/shared/dataprovider/api/model/vendorReport.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TransactionTypeEnum } from './transactionTypeEnum';

export interface VendorReport { 
    Vendor?: string;
    Type?: TransactionTypeEnum;
    Channel?: string;
    Fees?: number;
    Gross?: number;
    ACFees?: number;
    OTFees?: number;
    Count?: number;
    Net?: number;
}